+++ spec/rubocop/cop/style/empty_lines_spec.rb
@@ -19,7 +19,7 @@ describe RuboCop::Cop::Style::EmptyLines do
 
   it 'works when there are no tokens' do
     inspect_source(cop,
-                   ['#comment'])
+                   '#comment')
     expect(cop.offenses).to be_empty
   end
 
@@ -30,11 +30,11 @@ describe RuboCop::Cop::Style::EmptyLines do
   end
 
   it 'does not register an offense for empty lines in a string' do
-    inspect_source(cop, ['result = "test
+    inspect_source(cop, 'result = "test
 
 
 
-                                  string"'])
+                                  string"')
     expect(cop.offenses).to be_empty
   end